Refactored: Made map into gapMap class
